moved unit tests namespace - moved namespace of ControlLine::SendOperation() tests
