+++ HardwareComs.Tests.ControlLine.Unit/ControlLineSockets/SendOperationTests/Scenarios/Shared/Socket/When_Socket_Communication_Was_Successful.cs
@@ -2,7 +2,7 @@
 using NSubstitute;
 using NUnit.Framework;
 
-namespace ControlLineUnitTests.ControlLineSocketsTests.SendOperation.Scenarios.Shared.Socket
+namespace ControlLineUnitTests.ControlLineSockets.SendOperationTests.Scenarios.Shared.Socket
 {
     public abstract class When_Socket_Communication_Was_Successful : When_Preconditions_Are_Satisfied
     {
