+++ HardwareComs.Tests.ControlLine.Unit/ControlLineSockets/SendOperationTests/Scenarios/When_Connection_Cannot_Be_Opened.cs
@@ -1,10 +1,10 @@
 ﻿using System.Net.Sockets;
 using ControlLine.Dto;
-using ControlLineUnitTests.ControlLineSocketsTests.SendOperation.Scenarios.Shared.Socket;
+using ControlLineUnitTests.ControlLineSockets.SendOperationTests.Scenarios.Shared.Socket;
 using NSubstitute;
 using NUnit.Framework;
 
-namespace ControlLineUnitTests.ControlLineSocketsTests.SendOperation.Scenarios
+namespace ControlLineUnitTests.ControlLineSockets.SendOperationTests.Scenarios
 {
     [TestFixture]
     public class When_Connection_Cannot_Be_Opened : When_Socket_Error_Occurs
